Repackage the module files into `io.spine.web`.
