--- firebase-web/src/main/java/com/teamdev/licensing/web/firebase/package-info.java
@@ -1,14 +1,0 @@
-/*
- * Copyright (c) 2000-2018 TeamDev Ltd. All rights reserved.
- * TeamDev PROPRIETARY and CONFIDENTIAL.
- * Use is subject to license terms.
- */
-
-/**
- * This package contains the Firebase client transport.
- */
-
-@ParametersAreNonnullByDefault
-package com.teamdev.licensing.web.firebase;
-
-import javax.annotation.ParametersAreNonnullByDefault;
