+++ firebase-web/src/main/java/io/spine/web/firebase/package-info.java
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2000-2018 TeamDev Ltd. All rights reserved.
+ * TeamDev PROPRIETARY and CONFIDENTIAL.
+ * Use is subject to license terms.
+ */
+
+/**
+ * This package contains the Firebase client transport.
+ */
+
+@ParametersAreNonnullByDefault
+package io.spine.web.firebase;
+
+import javax.annotation.ParametersAreNonnullByDefault;
