--- web/src/main/java/com/teamdev/licensing/web/queryservice/package-info.java
@@ -1,15 +1,0 @@
-/*
- * Copyright (c) 2000-2018 TeamDev Ltd. All rights reserved.
- * TeamDev PROPRIETARY and CONFIDENTIAL.
- * Use is subject to license terms.
- */
-
-/**
- * This package contains the {@link com.teamdev.licensing.web.queryservice.AsyncQueryService
- * AsyncQueryService} interface and its implementations.
- */
-
-@ParametersAreNonnullByDefault
-package com.teamdev.licensing.web.queryservice;
-
-import javax.annotation.ParametersAreNonnullByDefault;
