+++ web/src/main/java/io/spine/web/package-info.java
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2000-2018 TeamDev Ltd. All rights reserved.
+ * TeamDev PROPRIETARY and CONFIDENTIAL.
+ * Use is subject to license terms.
+ */
+
+/**
+ * This package contains the basic client transport API.
+ */
+
+@ParametersAreNonnullByDefault
+package io.spine.web;
+
+import javax.annotation.ParametersAreNonnullByDefault;
