+++ web/src/main/java/io/spine/web/parser/package-info.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2000-2018 TeamDev Ltd. All rights reserved.
+ * TeamDev PROPRIETARY and CONFIDENTIAL.
+ * Use is subject to license terms.
+ */
+
+/**
+ * This package contains components responsible for parsing messages sent over the Web protocols.
+ */
+@ParametersAreNonnullByDefault
+package io.spine.web.parser;
+
+import javax.annotation.ParametersAreNonnullByDefault;
