+++ web/src/main/java/io/spine/web/queryservice/package-info.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2000-2018 TeamDev Ltd. All rights reserved.
+ * TeamDev PROPRIETARY and CONFIDENTIAL.
+ * Use is subject to license terms.
+ */
+
+/**
+ * This package contains the {@link io.spine.web.queryservice.AsyncQueryService
+ * AsyncQueryService} interface and its implementations.
+ */
+
+@ParametersAreNonnullByDefault
+package io.spine.web.queryservice;
+
+import javax.annotation.ParametersAreNonnullByDefault;
